+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## ChangeLog
 
+* The SDK will now check metrics for infinity and NaN.  Metrics with invalid
+values will be rejected, and will result in an error logged.
+
+* Added `Config.Product` and `Config.ProductVersion` fields which are
+used to the `User-Agent` header if set.
+
 ## 0.1.0
 
 First release!